Cleanup: merge variable/array access, remove param/var distinction
We simply model a variable access in the AST as an array access with the
index expression set to NULL. This allows us to merge some code. We can
still tell the two cases apart.

Because the declaration itself "knows" whether it is a function parameter
or an array, we can remove the distinction in the semantic analysis phase.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@student.uibk.ac.at>